+++ drivers/watty/device.ts
@@ -3,13 +3,13 @@ import _ from 'lodash';
 import moment from 'moment-timezone';
 import http from 'http.min';
 import { Subscription } from 'zen-observable-ts';
-import { LiveMeasurement, TibberApi, getRandomDelay } from '../../lib/tibber';
+import { LiveMeasurement, TibberApi } from '../../lib/api';
 import { NordPoolPriceResult } from '../../lib/types';
 import { startTransaction, noticeError } from '../../lib/newrelic-transaction';
 import { randomBetweenRange } from '../../lib/helpers';
 
 class WattyDevice extends Device {
-  #tibber!: TibberApi;
+  #api!: TibberApi;
   #deviceId!: string;
   #throttle!: number;
   #currency?: string;
@@ -37,7 +37,7 @@ class WattyDevice extends Device {
   async onInit() {
     const { id, t: token } = this.getData();
 
-    this.#tibber = new TibberApi(this.log, this.homey.settings, id, token);
+    this.#api = new TibberApi(this.log, this.homey.settings, id, token);
     this.#deviceId = id;
     this.#throttle = this.getSetting('pulse_throttle') || 30;
 
@@ -129,7 +129,7 @@ class WattyDevice extends Device {
 
     let websocketSubscriptionUrl = '';
     try {
-      const { viewer } = await this.#tibber.getHomeFeatures(this);
+      const { viewer } = await this.#api.getHomeFeatures(this);
       websocketSubscriptionUrl = viewer.websocketSubscriptionUrl;
 
       if (!viewer?.home?.features?.realTimeConsumptionEnabled) {
@@ -151,7 +151,7 @@ class WattyDevice extends Device {
 
     this.log('Subscribing to live data for homeId', this.#deviceId);
 
-    this.#wsSubscription = this.#tibber
+    this.#wsSubscription = this.#api
       .subscribeToLive(websocketSubscriptionUrl)
       .subscribe(
         (result) => this.subscribeCallback(result),
